Add Baikal map (fixed) (#690)
## Description:
This PR adds the "Baikal" map which is a cutout from the whole real life
lake, as the lake itself is very large vertically and wouldn't make as a
very good map by itself. This map has some rivers and a very big sort of
water body that seperates the map in two(it was kind of designed for the
teams gamemode). It's nations are some features around that region.

## Please complete the following:

- [ ] I have added screenshots for all UI updates
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced
- [x] I understand that submitting code with bugs that could have been
caught through manual testing blocks releases and new features for all
contributors

## Please put your Discord username so you can be contacted if a bug or
regression is found:

Nikola123