Ensuring that MutableAlliance::extend is called in test case (#1582)
## Description:

A test case for AllianceExtensionExecution was not actually checking if
the alliance was extended correctly. The method to extend the alliance,
MutableAlliance::extend, was never called during execution of
AllianceExtensionExecution.

The reason the extend() method was not being called is because both mock
players were not alive. This PR spies on the MutableAlliance object to
verify that extend() is correctly called.

Related: https://github.com/openfrontio/OpenFrontIO/pull/1536

## Please complete the following:

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText()
and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced
- [x] I have read and accepted the CLA agreement (only required once).

## Please put your Discord username so you can be contacted if a bug or
regression is found:

bypie5